+++ Beam/resources/assets/js/remplib.js
@@ -448,7 +448,7 @@ remplib = typeof(remplib) === 'undefined' ? {} : remplib;
                     "funnel_id": funnelId
                 },
                 "remp_commerce_id": remplib.uuidv4(),
-                "remp_checkout_id": remplib.generateRempCheckoutID(),
+                "commerce_session_id": remplib.generateCommerceSessionID(),
             };
             params = this.addSystemUserParams(params);
             this.post(this.url + "/track/commerce", params);
@@ -467,7 +467,7 @@ remplib = typeof(remplib) === 'undefined' ? {} : remplib;
                     "source": source
                 },
                 "remp_commerce_id": remplib.uuidv4(),
-                "remp_checkout_id": remplib.generateRempCheckoutID(),
+                "commerce_session_id": remplib.generateCommerceSessionID(),
             };
 
             params = this.addSystemUserParams(params);
@@ -489,7 +489,7 @@ remplib = typeof(remplib) === 'undefined' ? {} : remplib;
                     "product_ids": productIds
                 },
                 "remp_commerce_id": remplib.uuidv4(),
-                "remp_checkout_id": remplib.getRempCheckoutID(),
+                "commerce_session_id": remplib.getCommerceSessionID(),
             };
             params = this.addSystemUserParams(params);
             this.post(this.url + "/track/commerce", params);
@@ -512,7 +512,7 @@ remplib = typeof(remplib) === 'undefined' ? {} : remplib;
                     "source": source
                 },
                 "remp_commerce_id": remplib.uuidv4(),
-                "remp_checkout_id": remplib.getRempCheckoutID(),
+                "commerce_session_id": remplib.getCommerceSessionID(),
             };
 
             params = this.addSystemUserParams(params);
@@ -534,7 +534,7 @@ remplib = typeof(remplib) === 'undefined' ? {} : remplib;
                     "product_ids": productIds
                 },
                 "remp_commerce_id": remplib.uuidv4(),
-                "remp_checkout_id": remplib.getRempCheckoutID(),
+                "commerce_session_id": remplib.getCommerceSessionID(),
             };
             params = this.addSystemUserParams(params);
             this.post(this.url + "/track/commerce", params);
@@ -557,7 +557,7 @@ remplib = typeof(remplib) === 'undefined' ? {} : remplib;
                     "source": source
                 },
                 "remp_commerce_id": remplib.uuidv4(),
-                "remp_checkout_id": remplib.getRempCheckoutID(),
+                "commerce_session_id": remplib.getCommerceSessionID(),
             };
 
             params = this.addSystemUserParams(params);
@@ -579,7 +579,7 @@ remplib = typeof(remplib) === 'undefined' ? {} : remplib;
                     "product_ids": productIds
                 },
                 "remp_commerce_id": remplib.uuidv4(),
-                "remp_checkout_id": remplib.getRempCheckoutID(),
+                "commerce_session_id": remplib.getCommerceSessionID(),
             };
             params = this.addSystemUserParams(params);
             this.post(this.url + "/track/commerce", params);
@@ -602,7 +602,7 @@ remplib = typeof(remplib) === 'undefined' ? {} : remplib;
                     "source": source
                 },
                 "remp_commerce_id": remplib.uuidv4(),
-                "remp_checkout_id": remplib.getRempCheckoutID(),
+                "commerce_session_id": remplib.getCommerceSessionID(),
             };
             params = this.addSystemUserParams(params);
             params["user"]["source"] = source;